rename DelayedRender component
