+++ src/components/DelayedRender.js
